+++ Changelog.md
@@ -12,6 +12,7 @@
   - Allow placing supports manually
 - Add anti-aliasing to slice preview
 - Reduce output file size (Reported by Jens Holländer)
+- Remote Print: Support SDCP V3.0.0 for newer printers
 
 ## v0.8.0 &mdash; June 30th, 2026